feat(#012): Create event details view and styles
Modified the actual event_view.html and views.py of the event entity.
Styled the view with event.css.
You have to create the event, then the participants and finally
associate them to the previous event. Finally, you can go to
localhost:8000/evento/:id in order to see the view.

<footer>