test(#42): Add Login Selenium Test
These tests right now need the server to be running at the same time that the 'python manage.py test' command is done. I have disabled the option for them to be executed when that command is done so that the GitHub Actions do not fail. It would be necessary to add 'TestCase' in the class 'TestLogin'.