feat(#85): Add docker.yml file to configure continuous Docker deployment
Change the name of the main workflow of the project, from 'Innosoft eventos GA' to 'INNOSOFT eventos Build' in the main.yml file.
Add a new workflow in the docker.yml file for continuous deployment in docker, which depends on the execution of the main workflow.
Change the main workflow badge identifier in README from 'Innosoft eventos GitHub Actions' to 'INNOSOFT eventos Build'.